Get the mime types of the possible outputs for a report

Get the mime types of the possible outputs for a report

Usage

report_info(file)

Arguments

file

The path to the report

Value

A list with mime types of output and acceptable parameters