routr: A Simple Router for HTTP and WebSocket Requests

In order to make sure that web request ends up in the correct handler function a router is often used. 'routr' is a package implementing a simple but powerful routing functionality for R based servers. It is a fully functional 'fiery' plugin, but can also be used with other 'httpuv' based servers.

See also

Author

Maintainer: Thomas Lin Pedersen thomasp85@gmail.com (ORCID)